requirement.txt
altgraph==0.17.3
asgiref==3.5.2
autopep8==1.7.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
comtypes==1.1.14
cx-Freeze==6.13.0
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
djangorestframework==3.14.0
dnspython==2.2.1
filelock==3.8.0
future==0.18.2
gunicorn==20.1.0
idna==3.4
lief==0.12.2
lxml==4.9.1
MouseInfo==0.1.3
numpy==1.23.4
packaging==21.3
pandas==1.5.0
pefile==2022.5.30
Pillow==9.3.0
platformdirs==2.5.2
psycopg2==2.9.5
PyAudio==0.2.12
PyAutoGUI==0.9.53
pycodestyle==2.9.1
PyGetWindow==0.0.9
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.11
pymongo==4.2.0
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt6==6.4.0
PyQt6-Qt6==6.4.0
PyQt6-sip==13.4.0
PyRect==0.2.0
PyScreeze==0.1.28
PySimpleGUI==4.60.4
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.4
pywin32==304
pywin32-ctypes==0.2.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.5
urllib3==1.26.12
virtualenv==20.16.5
whitenoise==6.2.0
